Completed user: controller, services, model

+++ src/services/user/user.services.ts
@@ -2,11 +2,15 @@ import type { ROLES } from '../../constants/roles'
 import type { IUser } from '../../models/auth/auth-user.model'
 import {
 	deleteModelUser,
+	getModelUserById,
 	getModelUsers,
 	updatedModelUser,
 } from '../../models/users/user.model'
 import { UserDTO } from '../../utils/dtos/user-dto.utils'
 import { ApiError } from '../../utils/exists-error.utils'
+import logger from '../../utils/logger.utils'
+
+// !Todo: Make Controller for user  remove
 
 export interface IUserService {
 	user: UserDTO
@@ -16,30 +20,45 @@ class UserServices {
 	async getAllUsers(): Promise<IUser[]> {
 		const users = await getModelUsers()
 
-		if (!users) {
-			throw ApiError.BadRequest('Пользователи не были найдены')
+		if (!users?.length) {
+			throw ApiError.BadRequest('Пользователи не найдены')
 		}
 
 		return users
 	}
 
+	async getUserById(userId: number | string): Promise<IUser | null> {
+		return await getModelUserById(userId)
+	}
+
 	async updatedUser(
 		roleId: ROLES,
 		userId: number | string
-	): Promise<IUserService> {
+	): Promise<IUserService | null> {
 		const updatedUserData = await updatedModelUser(roleId, userId)
-
 		if (!updatedUserData) {
-			throw ApiError.BadRequest('Данный о user не обновились')
+			return null
 		}
 
-		const userDTO = new UserDTO(updatedUserData)
-
-		return { user: userDTO }
+		return { user: new UserDTO(updatedUserData) }
 	}
 
-	async removeUser(userId: string | number): Promise<void> {
-		await deleteModelUser(userId)
+	async removeUser(userId: string | number): Promise<IUser | null> {
+		if (!userId) {
+			logger.warn('Attempted to delete a user without ID')
+			throw ApiError.BadRequest('User ID is required')
+		}
+
+		const existingUser = await getModelUserById(userId)
+		if (!existingUser) {
+			logger.warn(`User ${userId} does not exist`)
+			return null
+		}
+
+		const removedUser = await deleteModelUser(userId)
+
+		logger.info(`User ${userId} deleted successfully`)
+		return removedUser
 	}
 }
